% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.CATS.R
\name{anova.CATS}
\alias{anova.CATS}
\title{Randomization-based Analysis of Deviance for CATS model}
\usage{
\method{anova}{CATS}(m, n.random = 49)
}
\arguments{
\item{m}{a fitted CATS model}

\item{n.random}{number of randomizations}
}
\value{
list of deviance tables
}
\description{
Calculates separate analysis of deviance tables for each plot (sample). P-values
are calculated by randomazation (reshuffling) of trait values. Correlations
among traits are conserved during randomization
}
