% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATSregression.R
\name{CATSregression}
\alias{CATSregression}
\title{Fitting Community Assembly via Trait Selection (CATS) model}
\usage{
CATSregression(
  abund,
  trait,
  prior = NULL,
  est.prior = FALSE,
  complexity = c("linear", "quadratic", "gam"),
  family = c("poisson", "binomial", "tw", "nb", "gaussian"),
  log.link = TRUE,
  trait.name = NULL,
  formul = NULL,
  n = NULL,
  ...
)
}
\arguments{
\item{abund}{matrix or data frame of abundances (species should be in columns)}

\item{trait}{vector of single trait or data frame of several traits}

\item{prior}{vector of meta-community level relative abundances}

\item{est.prior}{Should meta-community level relative abundances be estimated from abundances}

\item{complexity}{Predefined model formulas (see Details)}

\item{family}{Distribution of abundances}

\item{log.link}{Should log-link be used for binomial data instead of the cannonical logit?}

\item{trait.name}{Optional name of trait, if \code{trait} is a vector. If missing, the name will be "trait1"}

\item{formul}{Right-side of the formula to be fitted}

\item{n}{number of trials in the case of binomial data}
}
\value{
\code{CATSregression} returns an object of class CATS, which is a list containing the following components
\itemize{
        \item \code{params} list of input parameters
        \item \code{abund} observed abundances
        \item \code{model} list of objects of class gam or logbin
        \item \code{fitted} fitted abundances
        \item \code{R.sq} vector of plotwise goodness-of-fits measured by Kullback-Leibler R-squared
        \item \code{R.sq.adj} vector of plotwise goodness-of-fits measured by adjusted Kullback-Leibler R-squared
        \item \code{overall.R.sq} goodness-of-fit for the whole dataset measured by Kullback-Leibler R-squared
        \item \code{R.sq.adj} goodness-of-fit for the whole dataset measured by of adjusted Kullback-Leibler R-squared
        }
}
\description{
This function fits CATS model by generalized version of the procedure described by Warton et al. (2015)
}
\details{
This is the main function of the package. It fits the CATS model. As Warton et al.
(2015) has shown fitting GLM with Poisson distribution results in the same parameter
estimates as fitting model by the original maximum entropy approach. Advantage
of this approach is its higher flexibility: it allows changing distribution
(e.g. using negative binomial instead of Poisson for over-dispersed counts) and
modelling non-linear relationships by GAM instead of GLM.

In the recent version six distributions are supported:
\itemize{
   \item Poisson (family="poisson")
   \item binomial (family="binomial")
   \item negative binomial (family="nb")
   \item Tweedie with 1<power<2 (family="tw")
   \item normal (family="gaussian")
        }

\bold{Specifying model formula}
The fitted model should be specified either in parameter \code{formul} or in
\code{complexity}. The later allow three options:
\itemize{
   \item linear using all traits as predictor; using Poisson distribution it equivalent to
   fitting the original CATS model
   \item quadratic it includes all pairwise interactions (including quadratic terms)
   into the model
   \item gam it applied smoothing for each trait separatelyi
        }
}
\references{
Warton, D. I.; Shipley, B. & Hastie, T. (2015) CATS regression –
a model‐based approach to studying trait‐based community assembly
\emph{Methods in Ecology and Evolution} \bold{6}(4): 389-398.
\url{http://dx.doi.org/10.1111/2041-210X.12280}
}
