% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCATS.R
\name{checkCATS}
\alias{checkCATS}
\title{Cheking CATS model by parametric bootstrap}
\usage{
checkCATS(m, nrand = 100, draw.plot = TRUE, save.R2 = FALSE)
}
\arguments{
\item{m}{a fitted CATS model}

\item{nrand}{number of bootstrap samples}

\item{draw.plot}{Is a diagnostic plot drawn?}

\item{save.R2}{Should matrix of R-squared values be get as output?}
}
\value{
if \code{save.R2==T} matrix of R-squared values in bootsrap samples
(plots in colunms, bootstrap samples in rows); otherwise \code{NULL}
}
\description{
This function checks if the fitted model can describe the real assembly process.
}
\details{
The parametric bootstrap approach is applied: random samples are  generated
assuming that the fitted model describes accurately the real data. Then fit of
the original model to these parametric bootstrap data is measured by Kullback-Leibler
R-squared. If the model is accurate, the R-squared of real data will not differ
significantly from distribution R-squared of bootstrap samples.
}
