% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.CATS.r
\name{plot.CATS}
\alias{plot.CATS}
\title{Diagnostic plots}
\usage{
\method{plot}{CATS}(m, ...)
}
\arguments{
\item{m}{a fitted CATS model}

\item{...}{additional \code{\link[base]{graphical parameters}} passed to plot function}
}
\description{
It creates two diagnotic plots: (1) QQ-plot of residuals can be used for checking
normality of residuals. (2) Residuals vs predicted abundances allows checking
constant dispersion (homoscedasticity) of residuals.
}
