% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.CATS.R
\name{predict.CATS}
\alias{predict.CATS}
\title{Predicted abundances}
\usage{
\method{predict}{CATS}(m)
}
\arguments{
\item{m}{a fitted CATS model}
}
\value{
matrix of predicted abundances (i.e. their expected value)
}
\description{
It gets predicted abundances in a matrix form
}
