% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.CATS.R
\name{residuals.CATS}
\alias{residuals.CATS}
\title{Dunn-Smith residuals of CATS model}
\usage{
\method{residuals}{CATS}(m)
}
\arguments{
\item{m}{a fitted CATS model}
}
\value{
matrix of residulas
}
\description{
Calculates Dunn-Smith residuals of a fitted CATS model.
}
\details{
Advantage of these residulas is that if model was specified correctly, they
follow standard normal distribution.
}
