% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.CATS.r
\name{summary.CATS}
\alias{summary.CATS}
\title{Predicted abundances}
\usage{
\method{summary}{CATS}(x)
}
\arguments{
\item{m}{a model fitted by CATSregression function}
}
\value{
predicted values in a numeric vector
}
\description{
It gets the abundances predicted by the fitted CATS model
}
